Account for Text Descenders
Titles including letters like "ygq" etc have bits that go "below the
line", or in font terms they are descenders that go below the baseline.

As a result, those titles ended up not lining up vertically with ones
that didn't.

I don't know if this is the most right way to fix it, but it's
definitely an improvement in my eyes!